att crud aluno

+++ Aluno.cs
@@ -2,22 +2,29 @@
 using System;
 using System.Globalization;
 
+namespace Aula01 { 
 class Aluno
 {
     static string cs = @"server=sql.freedb.tech;userid=freedb_Mateus;password=Nt8pd!Qz#UktKES;database=freedb_Design-Pattern";
 
-    static void Main(string[] args)
+    public void AlunoMenu()
     {
         while (true)
         {
+            Console.Clear();
+
             Console.WriteLine("\nSelecione uma operação:");
+
+            Console.WriteLine("\n---------- Aluno ----------\n");
+
             Console.WriteLine("1. Inserir aluno");
             Console.WriteLine("2. Listar alunos");
             Console.WriteLine("3. Atualizar aluno");
             Console.WriteLine("4. Deletar aluno");
-            Console.WriteLine("5. Sair");
+            Console.WriteLine("0. Voltar ao menu\n");
 
-            int opcao = int.Parse(Console.ReadLine());
+                Console.Write("Opcao: ");
+                int opcao = int.Parse(Console.ReadLine());
 
             switch (opcao)
             {
@@ -33,8 +40,7 @@ class Aluno
                 case 4:
                     DeletarAluno();
                     break;
-                case 5:
-                    Console.WriteLine("Saindo...");
+                case 0:
                     return;
                 default:
                     Console.WriteLine("Opção inválida. Por favor, escolha uma opção válida.");
@@ -53,6 +59,8 @@ class Aluno
             string nome;
             do
             {
+                Console.Clear();
+                Console.WriteLine("Inserir Aluno \n\n");
                 Console.WriteLine("*Digite o nome do aluno: ");
                 nome = Console.ReadLine();
 
@@ -174,10 +182,12 @@ class Aluno
         }
     }
 
-    static void ListarAlunos()
+    static Boolean ListarAlunos()
     {
         try
         {
+            Console.Clear();
+
             using var crud = new MySqlConnection(cs);
             crud.Open();
 
@@ -199,17 +209,22 @@ class Aluno
                     string genero = reader.GetString("Genero");
 
                     Console.WriteLine($"Nome: {nome}, RGM: {rgm}, Data de Nascimento: {dataNasc}, Curso: {curso}, Bolsista: {bolsista}, RG: {rg}, Gênero: {genero}");
+                    Console.WriteLine("Aperte ENTER para seguir");
+                    Console.ReadLine();   
                 }
-            }
+                    return true;
+                }
             else
             {
                 Console.WriteLine("Nenhum aluno cadastrado.");
+                    return false;
             }
         }
         catch (Exception ex)
         {
             Console.WriteLine($"Erro: {ex.Message}");
-        }
+            return false;
+            }
     }
 
 
@@ -217,133 +232,148 @@ class Aluno
     {
         try
         {
-            ListarAlunos();
+            Console.Clear();
+            Boolean retorno = ListarAlunos();
 
-            long? rgm = null;
-            do
-            {
-                Console.WriteLine("*Digite o RGM do aluno que deseja atualizar: ");
-                string input = Console.ReadLine();
+                if (retorno)
+                {
+                    long? rgm = null;
+                    do
+                    {
 
-                if (string.IsNullOrEmpty(input))
-                {
-                    Console.WriteLine("O campo RGM não pode estar vazio.");
-                }
-                else if (input.Length != 11)
-                {
-                    Console.WriteLine("O RGM deve ter exatamente 11 caracteres.");
-                }
-                else if (!long.TryParse(input, out long value))
-                {
-                    Console.WriteLine("Valor de RGM inválido.");
+                        Console.WriteLine("\n\nAtualizar Aluno \n\n");
+                        Console.WriteLine("*Digite o RGM do aluno que deseja atualizar: ");
+                        string input = Console.ReadLine();
+
+                        if (string.IsNullOrEmpty(input))
+                        {
+                            Console.WriteLine("O campo RGM não pode estar vazio.");
+                        }
+                        else if (input.Length != 11)
+                        {
+                            Console.WriteLine("O RGM deve ter exatamente 11 caracteres.");
+                        }
+                        else if (!long.TryParse(input, out long value))
+                        {
+                            Console.WriteLine("Valor de RGM inválido.");
+                        }
+                        else
+                        {
+                            rgm = value;
+                        }
+                    } while (!rgm.HasValue);
+
+                    using var crud = new MySqlConnection(cs);
+                    crud.Open();
+
+                    var cmd = new MySqlCommand();
+                    cmd.Connection = crud;
+
+                    string novoNome;
+                    do
+                    {
+                        Console.WriteLine("*Digite o nome do aluno: ");
+                        novoNome = Console.ReadLine();
+
+                        if (string.IsNullOrEmpty(novoNome))
+                        {
+                            Console.WriteLine("O campo Nome não pode estar vazio.");
+                        }
+                        else if (novoNome.Length > 50)
+                        {
+                            Console.WriteLine("O campo nome deve ter no máximo 50 caracteres.");
+                        }
+                    } while (string.IsNullOrEmpty(novoNome) || novoNome.Length > 50);
+
+                    string novoDataNasc;
+                    do
+                    {
+                        Console.WriteLine("*Digite a nova data de nascimento do aluno (YYYY-MM-DD): ");
+                        novoDataNasc = Console.ReadLine();
+
+                        if (!DateTime.TryParseExact(novoDataNasc, "yyyy-MM-dd", null, DateTimeStyles.None, out _))
+                        {
+                            Console.WriteLine("Formato de data inválido. Por favor, insira a data no formato YYYY-MM-DD.");
+                        }
+                    } while (!DateTime.TryParseExact(novoDataNasc, "yyyy-MM-dd", null, DateTimeStyles.None, out _));
+
+                    string novoCurso;
+                    do
+                    {
+                        Console.WriteLine("*Digite o curso do aluno: ");
+                        novoCurso = Console.ReadLine();
+
+                        if (string.IsNullOrEmpty(novoCurso))
+                        {
+                            Console.WriteLine("O campo curso não pode estar vazio.");
+                        }
+                        else if (novoCurso.Length > 100)
+                        {
+                            Console.WriteLine("O campo curso deve ter no máximo 100 caracteres.");
+                        }
+                    } while (string.IsNullOrEmpty(novoCurso) || novoCurso.Length > 100);
+
+                    string novoBolsista;
+                    do
+                    {
+                        Console.WriteLine("*O aluno é bolsista? (1 para sim, 0 para não): ");
+                        novoBolsista = Console.ReadLine();
+
+                        if (novoBolsista.Length > 1)
+                        {
+                            Console.WriteLine("O campo bolsista deve ter no máximo 1 caracteres.");
+                        }
+                    } while (novoBolsista.Length > 1);
+
+                    long? novoRG = null;
+                    do
+                    {
+                        Console.WriteLine("*Digite o rg do aluno (somente numeros): ");
+                        string input = Console.ReadLine();
+
+                        if (string.IsNullOrEmpty(input))
+                        {
+                            Console.WriteLine("O campo rg não pode estar vazio.");
+                        }
+                        else if (input.Length > 11)
+                        {
+                            Console.WriteLine("O rg deve ter 11 caracteres ou menos.");
+                        }
+                        else if (!long.TryParse(input, out long value))
+                        {
+                            Console.WriteLine("Valor de rg inválido.");
+                        }
+                        else
+                        {
+                            novoRG = value;
+                        }
+                    } while (!novoRG.HasValue);
+
+                    Console.WriteLine("*Digite o gênero do aluno: ");
+                    string novoGenero = Console.ReadLine();
+
+                    cmd.CommandText = $"UPDATE Aluno SET Nome = '{novoNome}', DataNasc = '{novoDataNasc}', Curso = '{novoCurso}', Bolsista = {novoBolsista}, RG = {novoRG}, Genero = '{novoGenero}' WHERE Rgm = {rgm}";
+
+                    int rowsAffected = cmd.ExecuteNonQuery();
+
+                    if (rowsAffected > 0)
+                    {
+                        Console.WriteLine("Aluno atualizado com sucesso!");
+                    }
+                    else
+                    {
+                        Console.WriteLine("Aluno não encontrado.");
+                    }
                 }
                 else
                 {
-                    rgm = value;
+                    Console.WriteLine("\nNão há alunos para atualizar !\n\nAperte ENTER para voltar");
+                    Console.ReadLine();
+
                 }
-            } while (!rgm.HasValue);
 
-            using var crud = new MySqlConnection(cs);
-            crud.Open();
 
-            var cmd = new MySqlCommand();
-            cmd.Connection = crud;
-
-            string novoNome;
-            do
-            {
-                Console.WriteLine("*Digite o nome do aluno: ");
-                novoNome = Console.ReadLine();
-
-                if (string.IsNullOrEmpty(novoNome))
-                {
-                    Console.WriteLine("O campo Nome não pode estar vazio.");
-                }
-                else if (novoNome.Length > 50)
-                {
-                    Console.WriteLine("O campo nome deve ter no máximo 50 caracteres.");
-                }
-            } while (string.IsNullOrEmpty(novoNome) || novoNome.Length > 50);
-            
-            string novoDataNasc;
-            do
-            {
-                Console.WriteLine("*Digite a nova data de nascimento do aluno (YYYY-MM-DD): ");
-                novoDataNasc = Console.ReadLine();
-
-                if (!DateTime.TryParseExact(novoDataNasc, "yyyy-MM-dd", null, DateTimeStyles.None, out _))
-                {
-                    Console.WriteLine("Formato de data inválido. Por favor, insira a data no formato YYYY-MM-DD.");
-                }
-            } while (!DateTime.TryParseExact(novoDataNasc, "yyyy-MM-dd", null, DateTimeStyles.None, out _));
-
-            string novoCurso;
-            do
-            {
-                Console.WriteLine("*Digite o curso do aluno: ");
-                novoCurso = Console.ReadLine();
-
-                if (string.IsNullOrEmpty(novoCurso))
-                {
-                    Console.WriteLine("O campo curso não pode estar vazio.");
-                }else if(novoCurso.Length > 100 )
-                {
-                    Console.WriteLine("O campo curso deve ter no máximo 100 caracteres.");
-                }
-            } while (string.IsNullOrEmpty(novoCurso) || novoCurso.Length > 100);
-
-            string novoBolsista;
-            do
-            {
-                Console.WriteLine("*O aluno é bolsista? (1 para sim, 0 para não): ");
-                novoBolsista = Console.ReadLine();
-
-                if(novoBolsista.Length > 1 )
-                {
-                    Console.WriteLine("O campo bolsista deve ter no máximo 1 caracteres.");
-                }
-            } while (novoBolsista.Length > 1);
-
-            long? novoRG = null;
-            do
-            {
-                Console.WriteLine("*Digite o rg do aluno (somente numeros): ");
-                string input = Console.ReadLine();
-
-                if (string.IsNullOrEmpty(input))
-                {
-                    Console.WriteLine("O campo rg não pode estar vazio.");
-                }
-                else if (input.Length > 11)
-                {
-                    Console.WriteLine("O rg deve ter 11 caracteres ou menos.");
-                }
-                else if (!long.TryParse(input, out long value))
-                {
-                    Console.WriteLine("Valor de rg inválido.");
-                }
-                else
-                {
-                    novoRG = value;
-                }
-            } while (!novoRG.HasValue);
-
-            Console.WriteLine("*Digite o gênero do aluno: ");
-            string novoGenero = Console.ReadLine();
-
-            cmd.CommandText = $"UPDATE Aluno SET Nome = '{novoNome}', DataNasc = '{novoDataNasc}', Curso = '{novoCurso}', Bolsista = {novoBolsista}, RG = {novoRG}, Genero = '{novoGenero}' WHERE Rgm = {rgm}";
-
-            int rowsAffected = cmd.ExecuteNonQuery();
-
-            if (rowsAffected > 0)
-            {
-                Console.WriteLine("Aluno atualizado com sucesso!");
             }
-            else
-            {
-                Console.WriteLine("Aluno não encontrado.");
-            }
-        }
         catch (Exception ex)
         {
             Console.WriteLine($"Erro: {ex.Message}");
@@ -354,33 +384,47 @@ class Aluno
     {
         try
         {
-            ListarAlunos();
+            Console.Clear();
+            Boolean retorno = ListarAlunos();
+                if (retorno)
+                {
+                    Console.WriteLine("\n\nDeletar Aluno \n\n ");
+                    Console.WriteLine("Digite o RGM do aluno que deseja deletar: ");
+                    long rgm = long.Parse(Console.ReadLine());
 
-            Console.WriteLine("Digite o RGM do aluno que deseja deletar: ");
-            long rgm = long.Parse(Console.ReadLine());
+                    using var crud = new MySqlConnection(cs);
+                    crud.Open();
 
-            using var crud = new MySqlConnection(cs);
-            crud.Open();
+                    var cmd = new MySqlCommand();
+                    cmd.Connection = crud;
 
-            var cmd = new MySqlCommand();
-            cmd.Connection = crud;
+                    cmd.CommandText = $"DELETE FROM Aluno WHERE Rgm = {rgm}";
 
-            cmd.CommandText = $"DELETE FROM Aluno WHERE Rgm = {rgm}";
+                    int rowsAffected = cmd.ExecuteNonQuery();
 
-            int rowsAffected = cmd.ExecuteNonQuery();
+                    if (rowsAffected > 0)
+                    {
+                        Console.WriteLine("Aluno deletado com sucesso!");
+                    }
+                    else
+                    {
+                        Console.WriteLine("Aluno não encontrado.");
+                    }
 
-            if (rowsAffected > 0)
-            {
-                Console.WriteLine("Aluno deletado com sucesso!");
+                }
+                else
+                {
+                    Console.WriteLine("\nNão há alunos para deletar !\n\nAperte ENTER para voltar");
+                    Console.ReadLine();
+
+                }
+
+
             }
-            else
-            {
-                Console.WriteLine("Aluno não encontrado.");
-            }
-        }
         catch (Exception ex)
         {
             Console.WriteLine($"Erro: {ex.Message}");
         }
     }
+}
 }
